+++ src/datasources/iex.datasource.ts
@@ -1,15 +1,10 @@
 import { DataSource } from "apollo-datasource";
-import { company, earningsToday, keyStats, previousDay, price, quote } from "iexcloud_api_wrapper";
+import { company, earningsToday, keyStats, previousDay, quote } from "iexcloud_api_wrapper";
 
 export class IexAPI extends DataSource {
 	public getCompany = company;
 	public getEarningsToday = earningsToday;
 	public getKeyStats = keyStats;
 	public getPreviousDayPrice = previousDay;
-	public getPrice = price;
 	public getQuote = quote;
-
-	constructor() {
-		super();
-	}
 }