+++ src/graphql/resolvers/connection.resolver.ts
@@ -10,13 +10,14 @@ import {
 } from "@/graphql/pagination";
 import { models } from "@/mongodb";
 import { Maybe } from "@/types";
+import { doesExist } from "@/utils";
 import Base64URL from "base64-url";
 import DataLoader from "dataloader";
 import { GraphQLScalarType, Kind, ValueNode } from "graphql";
 import { IResolverObject, IResolvers } from "graphql-tools";
 import { Cursor as MongoDBCursor } from "mongodb";
 import { Document } from "mongoose";
-import { complement, indexBy, isNil, prop } from "ramda";
+import { indexBy, prop } from "ramda";
 
 export interface IConnectionArguments extends IPaginationParams {
 	before?: any;
@@ -56,8 +57,6 @@ export const ConnectionEdge: IResolverObject<Document, IServerContext> = {
 	node: (parent) => parent
 };
 
-const doesExist = <T>(value: T | null | undefined): value is T => complement(isNil)(value);
-
 const prime = <
 	S extends keyof typeof models,
 	T extends InstanceType<typeof models[S]> = InstanceType<typeof models[S]>
@@ -82,11 +81,9 @@ export const resolveRootConnection = async <
 	const {
 		connectors: { MongoDB }
 	} = context;
-	const { first, last, before, after, ...restArgs } = args;
+	const { first, last, before, after, ...filter } = args;
 
 	const source: IAbstractSourceWithCursor<T> = MongoDB.getWithCursor<T>(sourceName);
-
-	const filter: { [key: string]: any } = MongoDB.adaptQueryArgs(restArgs);
 
 	const query: IAbstractCursor<T> = await MongoDB.limitQueryWithId(source, filter, {
 		before,