+++ src/utils/general-functions.ts
@@ -1,0 +1,3 @@
+import { complement, isNil } from "ramda";
+
+export const doesExist = <T>(value: T | null | undefined): value is T => complement(isNil)(value);