+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.2] - 2019-11-02
+### Added
+- Added IEX Cloud data provider.
+- Added stock portfolio queries and mutations.
+- Added data key queries.
+- Added permissions layer to queries and mutaitons.
+- Added viewer query.
+
 ## 0.0.1 - 2019-10-28
 ### Added
 - Added authentication mutations.
@@ -18,4 +26,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed resources not stopping on completed request.
 
-[Unreleased]: https://github.com/mua-inc/app_server/compare/v0.0.1...HEAD
+[Unreleased]: https://github.com/mua-inc/app_server/compare/v0.0.2...HEAD
+[0.0.2]: https://github.com/mua-inc/app_server/compare/v0.0.1...v0.0.2